Basically rewrite the entire thing
It is not worth going into the details here, this has been coming for
a few years now.  Most of the effort has been done over the last few
days and most of it boils down to simplifications that avoid complex
state management.

The first apparent difference is that Ijod is now structured into
separate modules, making naming easier and nudging me to avoid global
state even more.

With bumping the minimal version of Go to 1.16, the need for
go-bindata falls away and is replaced by the build-in embed system.

Internally, the number of message types have been reduced.  The main
one now is the "state" type that shares the exact, time-stamped video
state a client finds himself in.  The latest state is always
remembered by the room daemon and shared with anyone entering the room
later on.  They can then use the time-stamp to deduce the actual time
stamp.

Notable features that have been dropped are

- The -wiki flag
- ACL management using .acl files (doesn't work anyway when using a
  reverse proxy)
- Reloading the directory tree on SIGUSR1 (the directory tree is not
- cached anymore, and generated anew each time it is requested)

All in all this warrants an unofficial "version 2", if Ijod were
actually versioned software.